+++ zapit/+zapit/+settings/private/default_settings.m
@@ -9,18 +9,22 @@ function settings = default_settings
     settings.NI.device_ID = 'Dev1';
     settings.NI.samplesPerSecond = 10000;
     settings.NI.AOrange = 10;
-    settings.NI.AOchans = 0:3;
     settings.NI.triggerChannel = 'PFI0';
+    settings.NI.wrapper = 'dotnet'; % 'vidrio' or 'dotnet'
 
     settings.scanners.voltsPerPixel = 2.2E-3;
+    settings.scanners.invertXscanner = 0;
+    settings.scanners.invertYscanner = 0;
 
     settings.laser.name = 'obis';
     settings.laser.laserMinMaxControlVolts = [1.5,3.5];
     settings.laser.laserMinMax_mW = [0,100];
-    settings.laser.maxValueInGUI = 20; %TODO not in validation scriptt
+    settings.laser.maxValueInGUI = 20; %TODO not in validation script
 
     settings.camera.default_exposure = 2000;
     settings.camera.micronsPerPixel = 19.3;
+    settings.camera.flipImageUD = 0;
+    settings.camera.flipImageLR = 0;
 
     settings.calibrateScanners.areaThreshold = 500;
     settings.calibrateScanners.calibration_power_mW = 10;
@@ -32,9 +36,10 @@ function settings = default_settings
 
     settings.experiment.defaultLaserFrequencyHz = 40; 
     settings.experiment.defaultLaserPowerMW = 5;
-    settings.experiment.maxStimPointsPerCondition = 2;
     settings.experiment.offRampDownDuration_ms = 250;
+    settings.experiment.maxStimPointsPerCondition = 2;
 
     settings.cache.ROI = [];
+    settings.cache.previouslyLoadedFiles = [];
 
 end % default_settings