+++ zapit/+zapit/@pointer/sendSamples.m
@@ -56,27 +56,6 @@ function varargout = sendSamples(obj, varargin)
     verbose = params.Results.verbose;
 
 
-
-    % If the user has specified an experiment directory path, we check whether a stimulus parameter
-    % log file exists there and make one if not. 
-    if logging && ~isempty(obj.experimentPath) && exist(obj.experimentPath,'dir')
-        d = dir(fullfile(obj.experimentPath,[obj.stimConfig.logFileStem,'*']));
-
-        if isempty(d)
-            % The user has defined an experiment directory and it does not contain a 
-            % stimulus parameter log file. We make one. 
-            logParamFname = obj.stimConfig.logStimulusParametersToFile(obj.experimentPath);
-            fprintf('Writing stimulus parameter log file to %s\n', ...
-                fullfile(obj.experimentPath,logParamFname))
-        end
-
-        % By this point there must be a parameter log file and, since we are logging, we write 
-        % a trial log file also. 
-        obj.stimConfig.logTrialToFile(obj.experimentPath, conditionNumber, laserOn, hardwareTriggered)
-
-    end
-
-
     % Choose a random condition if necessary
     if isempty(conditionNumber) || conditionNumber == -1
         r = randperm(obj.stimConfig.numConditions);
@@ -89,6 +68,27 @@ function varargout = sendSamples(obj, varargin)
         laserOn = r(1);
     end
 
+
+    % If the user has specified an experiment directory path, we check whether a stimulus parameter
+    % log file exists there and make one if not.
+    if logging && ~isempty(obj.experimentPath) && exist(obj.experimentPath,'dir')
+        d = dir(fullfile(obj.experimentPath,[obj.stimConfig.logFileStem,'*']));
+
+        if isempty(d)
+            % The user has defined an experiment directory and it does not contain a
+            % stimulus parameter log file. We make one.
+            logParamFname = obj.stimConfig.logStimulusParametersToFile(obj.experimentPath);
+            fprintf('Writing stimulus parameter log file to %s\n', ...
+                fullfile(obj.experimentPath,logParamFname))
+        end
+
+        % By this point there must be a parameter log file and, since we are logging, we write
+        % a trial log file also.
+        obj.stimConfig.logTrialToFile(obj.experimentPath, conditionNumber, laserOn, hardwareTriggered)
+
+    end
+
+
     if verbose
         fprintf('Stimulating area %d\n', conditionNumber)
     end
@@ -96,21 +96,24 @@ function varargout = sendSamples(obj, varargin)
 
     % Make the waveforms to play
     cs = obj.stimConfig.chanSamples;
-    waveforms = [cs.scan(:,:,conditionNumber), cs.light(:,:,conditionNumber)];
+    waveforms = cs(:,:,conditionNumber);
 
     % Disable laser if requested
     if laserOn == 0
         waveforms(:,3) = 0;
     end
 
-    % Connect only the first time this is called
-    if isempty(obj.DAQ.hAO) || ~strcmp(obj.DAQ.hAO.taskName,'sendSamples')
-        obj.DAQ.connectClockedAO('numSamplesPerChannel',size(waveforms,1), ...
-                                'hardwareTriggered', hardwareTriggered, ...
-                                'taskName','sendSamples');
-    end
+    % It will only connect if the existing task name is different
+    obj.DAQ.connectClockedAO('numSamplesPerChannel',size(waveforms,1), ...
+                            'hardwareTriggered', hardwareTriggered, ...
+                            'taskName','sendSamples');
   
     
+    % The current rampdown should be what is requested by this trial
+    obj.stimConfig.offRampDownDuration_ms = ...
+            obj.stimConfig.stimLocations(conditionNumber).Attributes.offRampDownDuration_ms;
+
+
     % write voltage samples onto the task
     obj.DAQ.writeAnalogData(waveforms);