+++ zapit/+zapit/@pointer/stopOptoStim.m
@@ -24,20 +24,12 @@ function stopOptoStim(obj, rampDownInMS)
     samplesPerSecond = obj.DAQ.samplesPerSecond;
     bufferSize = obj.DAQ.numSamplesInBuffer;
 
-    if isempty(bufferSize) || obj.DAQ.hAO.isTaskDone
+    if isempty(bufferSize) || obj.DAQ.isAOTaskDone
         return
     end
 
 
-    % If the user requests no ramp-down then we simply stop the task
-    if rampDownInMS ==0
-    % Zero everything
-        t = obj.DAQ.lastWaveform;
-        t(:) = 0;
-        obj.DAQ.writeAnalogData(t);
-        obj.DAQ.stop
-    end
-
+    % If the user requests no ramp-down or their ramp down duration is too short then we simply stop the task
 
     % Otherwise we do a ramp-down
     msPerBuffer = (bufferSize/samplesPerSecond) * 1E3;
@@ -49,6 +41,23 @@ function stopOptoStim(obj, rampDownInMS)
     end
 
     numBuffers = ceil(rampDownInMS / msPerBuffer);
+
+    if rampDownInMS == 0 || numBuffers == 1
+    % Zero everything
+         t = obj.DAQ.lastWaveform;
+        t(:) = 0;
+
+        % This loop is run three times to ensure we get the lines zeroed. This was determined by
+        % trial an error on an NI USB-6363. A PCIe might be different (TODO).
+        for ii=1:3
+            obj.DAQ.writeAnalogData(t);
+        end
+
+        obj.DAQ.stop
+        return
+    end
+
+
 
     % The series of amplitudes over which we will loop
     smoothRamp = false; % if true we ramp the waveform nicely and not in steps
@@ -69,7 +78,7 @@ function stopOptoStim(obj, rampDownInMS)
     if obj.simulated
         data = [];
     end
-    verbose = true; % For debugging the rampdown
+    verbose = false; % For debugging the rampdown
     if verbose
         fprintf('There are %d steps in the rampdown\n', length(ind))
     end
@@ -78,6 +87,7 @@ function stopOptoStim(obj, rampDownInMS)
     wave = {};
     n=1;
     % calculate the waveforms to play
+    ampSequence
     for ii = ind
         if smoothRamp
             endVal = ampSequence(ii);
@@ -94,8 +104,16 @@ function stopOptoStim(obj, rampDownInMS)
             data = [data;t]; %#ok<AGROW> 
         end
 
+        % Disable the masking light when we are on the last cycle.
+        % Unless this is done here there is a tendency for the masking
+        % light to remain on at the end.
+        if n <= length(ampSequence)
+            t(:,4) = 0;
+        end
+
         wave{n} = t;
         n = n+1;
+
     end
 
     % Dump all the waveforms to the DAQ
@@ -111,9 +129,14 @@ function stopOptoStim(obj, rampDownInMS)
         ylabel('Voltage')
     end
 
-    % Zero everything
+    % Zero everything. We might have to send the buffer more than once to get the laser to zero
+    % depending on the number of passes through the above loop. This was determined by
+    % trial an error on an NI USB-6363. A PCIe might be different (TODO).
+    minBuffers = 6;
     t(:) = 0;
-    obj.DAQ.writeAnalogData(t);
+    for ii = 1: (1+minBuffers-numBuffers)
+        obj.DAQ.writeAnalogData(t);
+    end
 
     obj.DAQ.stop